Tied in Login Functionality
Created a project library to use in the solution so that we can reuse code. Tied in login functionality to the desktop application
