+++ code/RecipePlannerApplication/ProjectLibrary/Login.cs
@@ -1,0 +1,5 @@
+public class Login
+{
+    public string? Username { get; set; }
+    public string? Password { get; set; }
+}
