+++ code/RecipePlannerApplication/RecipePlannerApplication.sln
@@ -4,8 +4,13 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 VisualStudioVersion = 17.4.33213.308
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RecipePlannerDesktopApplication", "RecipePlannerDesktopApplication\RecipePlannerDesktopApplication.csproj", "{64EAD739-A5C8-4243-9866-A640FCBBC89C}"
+	ProjectSection(ProjectDependencies) = postProject
+		{4EC23151-B6F5-4933-98C5-410D4451793D} = {4EC23151-B6F5-4933-98C5-410D4451793D}
+	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RecipePlannerWebApplication", "RecipePlannerWebApplication\RecipePlannerWebApplication.csproj", "{4EC23151-B6F5-4933-98C5-410D4451793D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RecipePlannerWebApplication", "RecipePlannerWebApplication\RecipePlannerWebApplication.csproj", "{4EC23151-B6F5-4933-98C5-410D4451793D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RecipePlannerLibrary", "RecipePlannerLibrary\RecipePlannerLibrary.csproj", "{EF498F02-1D03-4ED0-8223-8178DE010B8F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +26,10 @@ Global
 		{4EC23151-B6F5-4933-98C5-410D4451793D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4EC23151-B6F5-4933-98C5-410D4451793D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4EC23151-B6F5-4933-98C5-410D4451793D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EF498F02-1D03-4ED0-8223-8178DE010B8F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EF498F02-1D03-4ED0-8223-8178DE010B8F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EF498F02-1D03-4ED0-8223-8178DE010B8F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EF498F02-1D03-4ED0-8223-8178DE010B8F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
