+++ code/RecipePlannerApplication/RecipePlannerLibrary/Connection.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace RecipePlannerLibrary
+{
+    public class Connection
+    {
+        public static readonly string LoginConnectionString = "Server=localhost;Database=capstone;uid=root;pwd=passwordpassword123;";
+    }
+}
