+++ code/RecipePlannerApplication/RecipePlannerLibrary/Login.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace RecipePlannerLibrary
+{
+    public class Login
+    {
+        public string? Username { get; set; }
+        public string? Password { get; set; }
+    }
+}
