--- code/RecipePlannerApplication/RecipePlannerWebApplication/Models/Connection.cs
@@ -1,7 +1,0 @@
-﻿namespace RecipePlannerWebApplication.Models
-{
-    public class Connection
-    {
-        public static readonly string LoginConnectionString = "Server=localhost;Database=capstone;uid=root;pwd=Brandon.2018;";
-    }
-}
