--- code/RecipePlannerApplication/RecipePlannerWebApplication/Models/Login.cs
@@ -1,8 +1,0 @@
-﻿namespace RecipePlannerWebApplication.Models
-{
-    public class Login
-    {
-        public string? Username { get; set; }
-        public string? Password { get; set; }
-    }
-}
