+++ code/RecipePlannerApplication/RecipePlannerWebApplication/RecipePlannerWebApplication.csproj
@@ -11,4 +11,8 @@
     <PackageReference Include="System.Data.SqlClient" Version="4.8.5" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\RecipePlannerLibrary\RecipePlannerLibrary.csproj" />
+  </ItemGroup>
+
 </Project>
