+++ code/RecipePlannerApplication/RecipePlannerWebApplication/Views/Home/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model RecipePlannerWebApplication.Models.Login
+﻿@model RecipePlannerLibrary.Login
 @{
     ViewData["Title"] = "Home Page";
 }
